Rework serialization format to be XHTML
Trying to parse HTML using XML tools is hacky and prone to fail in
spectacular ways. And when it fails, we risk losing data, which is very
bad.
Instead, we can make WebKit give us XM and store it directly, using XHTML.
The downside of this is that we need to give WebKit the full document
(including <html xmlns="..."><body>...). The upside is that, as we no
longer parse the note content, anything that can be represented in XHTML
(including images, forms, scripts, plugins...) can now be stored in a
note.

https://bugzilla.gnome.org/show_bug.cgi?id=690860