Instead of pixbufs. This enables crisp rendering of overview thumbnails on hidpi displays.
This commit also removes the abilty to load and save icons to disk,
since we don't want to cache them while their appearance changes e.g.
due to the theme.
The only consequence for now is that we won't get full icons in the
search provider, but that will be fixed in a later commit.