manager: rework how the manager is constructed
We want to use the manager from the shell search provider, to be able to
load note thumbnails now that we don't save them to disk anymore.
Split it in sync and async constructors using GInitable, and have the
main application still use the async constructor. The search provider
wants all the information loaded upfront so it will use the synchronous
constructor.
Along the way, code was ported to use g_icon_serialize() and many leaks
were fixed.

https://bugzilla.gnome.org/show_bug.cgi?id=751683