bjb-bijiben: improved how files opening is handled
On one hand file paths were converted back and forth to GFiles, on
the other hand paths were removed and added to the files_to_open
queue.

The code has been refactorized to avoid this and the files_to_open
has been changed to a static variable to avoid a memory allocation.