note-obj: Convert webkit1 to webkit2 notes
After migration to webkit2 API, when user tried to change notes created
before the migration, those updates were lost.

The issue was related to how old notes were saved, using different html
tags for header and javascript enablement, necessary for the
webkit-editor window.

This code was created to detect a webkit1 note and convert it to
webkit2, so webkit-editor won't get lost.

https://bugzilla.gnome.org/show_bug.cgi?id=782986