editor-toolbar: Remove accelerators
When GtkBuilder is creating the composite template
instance, it parses the accelerators and attempts to
find the parent toplevel widget to map the accelerators.

The problem with this is that the toolbar widget is not
added to any toplevel on creation time, and the accel
fails to be attached to the non-existant toplevel. This
throws a few warnings that can be avoided, since the
accels are also set programmatically.

Fix these warnings by removing the GtkBuilder accel
definitions from editor-toolbar.ui.