main-toolbar: Stuck in Trash View
When you empty trash, in trash view, the main toolbar changes and you
get stuck in that view. The only way to leave it is closing the
application.

That happens because the window state changes from ARCHIVE_VIEW to
NO_NOTE, but this stated is shared by MAIN_VIEW as well, so it gets
lost.

This commit change the behavior to whenever you click empty trash, the
window state continues to be ARCHIVE_VIEW, which is the default when you
check Trash for the first time even without items.

https://bugzilla.gnome.org/show_bug.cgi?id=763971