window: don't be aggressive in saving geometry
window geometry was been saved to gsettings for every minor change.
This was causing too much I/O.

So, save the geometry only after the change is complete (100 ms  is
taken as the threshold for this).

Also, since configure-event signal is emitted on window maximized/minimize
events, avoid using window-state-event handlers.

https://bugzilla.gnome.org/show_bug.cgi?id=786664

Signed-off-by: Isaque Galdino <igaldino@gmail.com>