note-obj: Critical error when creating a new note
When creating a new note, Bijiben is thrown a message: "Glib critical
error when creating a new note".

That happens because when it's checking for old notes, it's also trying
to convert new notes, which are null at the start of the process.

This patches checks that and avoids that behaviour.

https://bugzilla.gnome.org/show_bug.cgi?id=787289