main-view: Crash when closing app with empty note
Application is crashing when closing it in an empty note.

When application is closed, it's raised an item-delete action. The
item-delete process tries to get back to main view, but view are now
destroyed, so it reaches a NULL pointer.

This patch checks if view is NULL and then returns without further
processing.

https://bugzilla.gnome.org/show_bug.cgi?id=787508