application: Port to bjb-application
The class isn't derivable, so there is no point of making it private.
Also use G_DECLARE_FINAL_TYPE to reduce lots of boilerplate code.
And to be consistent with gnome coding standard, rename the file
to bjb-application.[ch]

This is a part of effort to clean up codebase and make the code
more maintainable.
https://bugzilla.gnome.org/show_bug.cgi?id=789696