app-menu: Simplify setting Application menu
gtk_application() looks for "gtk/menus.ui" and set it as app-menu if
it has a menu with id "app-menu".

So set the resource path such so that we don't have to manually take
care of configuring the menu.

This is a part of effort to clean up codebase and make the code
more maintainable.
https://bugzilla.gnome.org/show_bug.cgi?id=789696