import-dialog: Use symbolic icons
Using symbolic icons will improve the UI, and help us remove
some old code that was looking for the application icons manually.

Let us not care whether those application is actually installed,
but only check if their data directory is present. Which means
that even the application (Gnote/Tomboy) got uninstalled, their
notes can be imported in Bijiben.

This is a part of effort to clean up codebase and make the code more
maintainable.
https://bugzilla.gnome.org/show_bug.cgi?id=789696