date-time: Rewrite logic to get date difference
Blindly getting week, month, and year based on number of days
won't work right.

Let's use GDate to deduce this the right way.

Also GTimeVal is always 32 bit. Which means that the application
will break on 2038. Let's fix this so that mission critical systems
running bijiben won't fail on the date. :-)

https://bugzilla.gnome.org/show_bug.cgi?id=790779