memo-provider: Don't add custom border to pixbuf
Adding a custom border with the pixbuf doesn't make any difference.
Also, by removing this, we can remove a further dependency on libgd.
This was also causing a segfault as some display related code was
being run by libgd in the non-UI thread.

This commit avoids the custom border along with removing the related
libgd code, and saving from a crash

https://bugzilla.gnome.org/show_bug.cgi?id=761338