own-cloud-note: Fix saving notes
Owncloud integration is buggy. Saving notes action happens multiple
times, mainly when changing notes titles in an async way. That triggers
many file update actions in paralell and, in case of title change,
causes notes files being lost in the process.

This patch makes all file saving actions to be in sync, to make sure
that all operations will happen in order, preventing loosing data.
That may increase the saving time due to operations being in sync, but
it also reduces the number of saving actions.

https://bugzilla.gnome.org/show_bug.cgi?id=786239