menu: Move last updated text to burger menu
The issue is explained in bug report #3, but basically the "Last
updated ..." text is shown above the notes text.

There were some fixes done in the past which moved the text to the right
side of the window but that still was not good enough as text could
easily span to the right side of the window as well.

So, Allan Day (@aday) and Cosimo Cecchi (@cosimoc) proposed to move the
"Last updated ..." text to the menu as an insensitive item.

The change was implemented by Carl van Tonder but it was never pushed to
master.

This commit is hugely based on Carl's work and modified to our current
baseline.

Carl, if you see this: thank you!