css: Improve font and paragraph
Migration to webkit 2 introduced a regression in paragraphs as they
don't have the space required by issue #39 anymore.

That's because in webkit 1 paragraphs were wrapped by two div tags, but
now in webkit 2 they are wrapped by only one.

This commit fix that and it also increases note title font size to be
just like it's shown in mockups from #52.