css: Fix note title style
Note title was not changing its color like other note's texts.

After some debuging, it was noticed in the generated HTML that every
paragraph was wrapped by a 'div' tag but the first one, which is
actually the note title.

Many other problems, related to the first line missing the 'div' tag,
was found while googling. No explanation about that behaviour was found
in the net either.

Considering that, the implemented solution applies the note title style
to the whole HTML document and then the regular note text style for all
div elements.