webkit-editor: Add Zoom Levels
It was requested by our GNOME Design Team [1] to create a menu entry for
users to select one out of three notes' text sizes.

In this second commit, we have added the code to implement zoom levels
in webkit editor.

That is a better and simpler approach than changing font sizes in the
editor, which would make code more complex to deal with notes with many
font sizes.

[1] https://gitlab.gnome.org/GNOME/gnome-notes/issues/30