flatpak: Workaround Tracker issue
Tracker is preventing Notes to start a connection when in a sandbox
environment - flatpak. That is only happening in master branch.

This commit changes Notes flatpak to use the last known working Tracker
version: 2.0.3.

There is an open issue [1] for Tracker team already and we'll revert
this commit once that's fixed.

[1] https://gitlab.gnome.org/GNOME/tracker/issues/26