app-menu: Revert removal
This commit adds back the application menu, which was removed in commit
3d9df51e and replaced by a primary (hamburger) menu in commit 38c8c37c.

The first proposal [1] suggested applications could remove application
menu in 3.30 cycle already, but then the goal was reviewed [2] and a new
timeline was set to 3.32 cycle instead.

This commit also adds the "Text Size" application menu option (#30),
which was not in the removed application menu, but it was added to the
hamburger menu in commit 23776cf3, so application menu will be in sync
with options already in primary menu in master branch.

[1] https://wiki.gnome.org/Design/Whiteboards/AppMenuMigration
[2] https://wiki.gnome.org/Initiatives/GnomeGoals/AppMenuRetirement