Finish renaming to org.gnome.Notes
This renames the icons, search provider, gsettings schema and
application ID.

The code still has lots of instances of "bijiben", either in function
names or GResource paths, but these are internal implementation details
and don't cause any problem.

The things changed in this commit should be enough to get a working
Flatpak application.

Relates to #95