webkit-editor: Migrate to JSC
This change started as a fix to a few obsolete warnings during
compilation time. Compiler was complainting about
webkit_javascript_result_get_global_context [1] and
webkit_javascript_result_get_value [2] being obsolete since WebkitGtk+
2.22 version.

After some [re]search, I came across this Epiphany commit [3], which
decribes how to migrate to JavaScriptCore (JSC) API. This commit is
based on the applicable changes [4] from that commit.

As a side effect of this change, the code is easier to read and to
understand.

[1] https://webkitgtk.org/reference/webkit2gtk/stable/WebKitWebView.html#webkit-javascript-result-get-global-context
[2] https://webkitgtk.org/reference/webkit2gtk/stable/WebKitWebView.html#webkit-javascript-result-get-value
[3] https://gitlab.gnome.org/Community/Purism/epiphany/commit/85f6037a8d6785b6b2a5a4bfd83ea595d35aeb52
[4] https://gitlab.gnome.org/Community/Purism/epiphany/commit/85f6037a8d6785b6b2a5a4bfd83ea595d35aeb52?view=inline#36d94ded0e587d24b4e4a9e43f6ff247cf7437d6_3185_3178