tracker: Make private store optional
In commit dd4fec2f we have changed the way Notes handle Tracker, to use
it as a library instead of a service.

That is cool for sandbox package where it doesn't need to rely on
anything else.

But it might be interesting, when installed as a standard application,
to use the system service instead of a local one.