shell-search-provider: Don't free icon
Notes shell search provider was crashing when freeing notes objects.

The reason was that in order to provide an icon for GNOME Shell,
application was getting note's icon, which is a cairo_surface_t object,
creating a GdkPixbuf from that, and finally freeing the cairo_surface_t
object.

The problem is that the cairo_surface_t object, AKA note's icon, was
still being referenced by the note object, so when the note object was
freed, it tried to freed cairo_surface_t object, which address was
already invalid.