note-obj: Fix all text in bold
When application was creating a note from a text, it was replacing all
"\n" character with a "<br/>" tab element.

This change fixes that, wrapping every text line up in a "<div></div>"
pair tag element, taking care to not do that for the first line, which
is the note title.

This change also fixes the CSS for the title.