providers: Fix creation of new notes
After the change in the way we display last updated information (commits
92257ffd and f7cc2aac), own/nextcloud notes started to mishaving not
being able to create notes anymore but crashing.

At first I thought it was related to the above change, but it was
actually an old issue in Notes code which due to the change it was more
visible.

That is related to the note timestamp (mtime) used when creating new
own/nextcloud or memo notes. The previous code was using g_get_real_time
function, which is fine, but it was not converting it from miliseconds
to seconds.

This change fixes that, adding the missing conversion.