own-cloud-note: Fix note opening after creation
Own/Nextcloud notes weren't being opened after it's creation. It was
necessary to restart application in order to open them.

When a new note is created, the manager (BijiManager) creates an id and
adds it to its internal list of notes. For local notes, this id is also
the notes filename, which doesn't change when notes are updated.

This issue was happening because Own/Nextcloud notes uses the note's
title as its filename and the manager was not aware of that.

This commit fixes the manager so it updates its internal list of notes
with the new Own/Nextcloud id.