controller: Fix missing notes
The sort process of the main view was missing some notes after a user
updated them.

The issue was found when we noticed the sorting was done on a temporary
object instead of the main one. g_list_sort might change the start of
the list during the sorting process and when that happened it changed
the temporary object. As the main object was not changed, it continued
to point to the same node as before, which then was the second one. So,
when the update view function used the main object, it started from the
second node, instead of the first one.

This commit fixes that, removing the temporary object from the process.