manager: Fix restore from trash
GNOME Notes keeps notes references as its path in the file system. That
is known internally as uuid.

When a note is moved to trash, although its path changes, to add the
".Trash" directory path, its uuid does not.

When GNOME Notes was restoring a note from trash it was using the new
path, with ".Trash" directory path in it, instead of the uuid.

This commit changes that, so GNOME Notes now only uses the uuid.