controller: Add display-notebooks-changed signal
Previously, signal dispaly-items-changed is used to notify that notes
or notebooks has changed in the controller.

Now we need to show notes and notebooks separately in different lists,
it's better to add a new signal named display-notebooks-changed
indicating that notebooks has changed. signal display-items-changed is
renamed as display-notes-changed to indicates notes has changed.