nextcloud-provider: change how to get user domain
In the past, the NextCloud user and domain(server) information were used
for display only.

With the change to call the NextCloud REST API directly, that
information is also used to authenticate the user with the cloud provider.

There was an issue when we had a '@' in the user id, e.g. user@email
because the code was extracting that from the GOA presentation identity
field which concats the user with the domain using a '@' as well, e.g.
user@email@domain.

When the code was split it to get user and domain, the information was
messed up with user being 'user' and domain being 'email@domain'.

Fix issue #171