Add BjbNotebooksDialog
This dialog shall be used to edit notebooks associated
with a note, replacing BjbOrganizeDialog.

GListModel is used for handling data instead of GtkTreeView
which is a lot easier to use and is going to help ease
migration to GTK4.

Also, the latest design has been implemented (not as such, as we
don't yet have notebook deletion/rename support)