item: Remove unused methods
And adapt to changes

After moving to the new design, a lot of code has been not used.
We no longer use pre-existing icons/previews since we now simply create
a row label and we no longer have Grid views which were the ones
that were using preview icons.

Also remove biji-item methods like has_color(), get_place(), is_collectable()
which were never used.

Remove them so that it's easier to make changes and it will help
split biji-notebook and biji-note-obj (and renaming it to biji-note)
and to remove biji-item without much hassle.