manager: Store notes and notebooks separately
Previously, a GListStore is used to store notebooks and a GHashTable
is used to store all notes and notebooks. This leads to an issue that
an empty note list instead of empty page is shown when there are no
notes available, but notebooks.

Fix that by using a GHashTable to only store notes so that the numbers
of notes available can be return when needed.