controller: Update note list view based on note counts
This is a combined effort with the previous commit to fix the issue
where an empty note list show instead of empty page when there are no
notes but notebooks. Besides, this also tries to fix the issue where
window view mode doesn't change when note count change from 0 to 1,
or 1 to 0.

Only save a list of notes in the controller to avoid counting notebooks
as notes. And in on_manager_changed(), update window view mode according
to the changes of notes count.