Move libbiji source to src
Currently the only need for libbiji (or the core source) is the
shell search provider.  The source of libbiji is already dispersed
in libbiji and src directories.  Move it to src directory so it's
easier to handle (for the time being).

Currently we are moving around a lot of code.  We shall move it to
better locations (if needed) once the dust settles.  And since this
doesn't rewrite git history much, it shouldn't be a big problem.